refactor(exec): move InferenceState to exec/inference_state.h
